28/pyproject.toml
[project]
name = "sd-28"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "dspy>=3.0.4",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
